+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/Engine.kt
@@ -115,6 +115,11 @@ interface Engine :
         MAX,
 
         /**
+         * DoH setting is set to "Ultra", corresponds to TRR_MODE_ONLY (3) with OHTTP enabled.
+         */
+        ULTRA,
+
+        /**
          * DoH is disabled, corresponds to TRR_MODE_DISABLED (5) from GeckoView.
          */
         OFF,