+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Core.kt
@@ -179,6 +179,7 @@ class Core(
             dohProviderUrl = context.components.settings.dohProviderUrl,
             dohDefaultProviderUrl = context.components.settings.dohDefaultProviderUrl,
             dohExceptionsList = context.components.settings.dohExceptionsList.toList(),
+            dohUseOhttp = context.components.settings.dohUseOhttp,
             globalPrivacyControlEnabled = context.components.settings.shouldEnableGlobalPrivacyControl,
             fdlibmMathEnabled = FxNimbus.features.fingerprintingProtection.value().fdlibmMath,
             emailTrackerBlockingPrivateBrowsing = true,